Calculate promotion for soup and bread

+++ src/main/java/com/henrysgrocery/Product.java
@@ -1,20 +1,22 @@
 package com.henrysgrocery;
 
+import java.math.BigDecimal;
+
 public enum Product {
-	SOUP("tin", 0.65),
-	BREAD("loaf", 0.80),
-	MILK("bottle", 1.30),
-	APPLE("single", 0.10);
+	SOUP("tin", BigDecimal.valueOf(0.65)),
+	BREAD("loaf", BigDecimal.valueOf(0.80)),
+	MILK("bottle", BigDecimal.valueOf(1.30)),
+	APPLE("single", BigDecimal.valueOf(0.10));
 
 	private final String unit;
-	private final double price;
+	private final BigDecimal price;
 
-	Product(String unit, double price) {
+	Product(String unit, BigDecimal price) {
 		this.unit = unit;
 		this.price = price;
 	}
 
-	Double getPrice(){
+	BigDecimal getPrice(){
 		return price;
 	}
 }
